test(plugin): express button journeys through domain helpers
